Put SCSS rules in common file to keep it DRY
Depending on the mode, we will append the correct loaders.
