+++ src/App/webpack.config.dev.js
@@ -5,44 +5,5 @@ module.exports = {
     devtool: "inline-source-map",
     plugins: [
         new webpack.NamedModulesPlugin(),
-    ],
-    module: {
-        rules: [
-            {
-                test: /\.scss$/,
-                use: [
-                    {
-                        // CSS is outputted in the JS file to support HMR
-                        loader: "style-loader",
-                    },
-                    {
-                        loader: "css-loader",
-                        options: {
-                            sourceMap: true,
-                        },
-                    },
-                    {
-                        loader: "postcss-loader",
-                        options: {
-                            plugins: function() {
-                                return [
-                                    require("autoprefixer"),
-                                ];
-                            },
-                            sourceMap: true,
-                        },
-                    },
-                    {
-                        loader: "sass-loader",
-                        options: {
-                            includePaths: [
-                                "./node_modules",
-                            ],
-                            sourceMap: true,
-                        },
-                    },
-                ],
-            },
-        ],
-    },
+    ]
 };
