+++ src/App/webpack.config.prod.js
@@ -36,51 +36,4 @@ module.exports = {
     plugins: [
         new CleanWebpackPlugin([targetDir])
     ],
-    module: {
-        rules: [
-            {
-                test: /\.scss$/,
-                use: [
-                    {
-                        loader: "file-loader",
-                        options: {
-                            name: '[name].css'
-                        }
-                    },
-                    {
-                        loader: "extract-loader",
-                        options: {
-                            publicPath: '/build/'
-                        }
-                    },
-                    {
-                        loader: "css-loader",
-                        options: {
-                            sourceMap: true,
-                        },
-                    },
-                    {
-                        loader: "postcss-loader",
-                        options: {
-                            plugins: function() {
-                                return [
-                                    require("autoprefixer"),
-                                ];
-                            },
-                            sourceMap: true,
-                        },
-                    },
-                    {
-                        loader: "sass-loader",
-                        options: {
-                            includePaths: [
-                                "./node_modules",
-                            ],
-                            sourceMap: true,
-                        },
-                    },
-                ],
-            },
-        ],
-    },
 };
